# docs/api-spec.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "RedTeam Automation Platform API",
    "description": "API for managing bug bounty programs and security automation workflows",
    "version": "1.0.0",
    "contact": {
      "name": "RedTeam Automation Support",
      "email": "support@redteam-automation.com"
    },
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    }
  },
  "servers": [
    {
      "url": "http://localhost:3001/api",
      "description": "Development server"
    },
    {
      "url": "https://api.redteam-automation.com",
      "description": "Production server"
    }
  ],
...
  "paths": {
    "/auth/register": {
      "post": {
        "tags": ["Authentication"],
        "summary": "Register a new user",
        "description": "Create a new user account with email and password",
        "operationId": "registerUser",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "User created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid input data",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "409": {
            "description": "Email already exists",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/auth/login": {
      "post": {
        "tags": ["Authentication"],
        "summary": "User login",
        "description": "Authenticate user with email and password",
        "operationId": "loginUser",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Login successful",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthResponse"
                }
              }
            },
            "headers": {
              "Set-Cookie": {
                "description": "HTTP-only cookie containing JWT token",
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "401": {
            "description": "Invalid credentials",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "429": {
            "description": "Too many login attempts",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/auth/logout": {
      "post": {
        "tags": ["Authentication"],
        "summary": "User logout",
        "description": "Logout current user and invalidate token",
        "operationId": "logoutUser",
        "security": [{"BearerAuth": []}],
        "responses": {
          "200": {
            "description": "Logout successful",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/programs": {
      "get": {
        "tags": ["Programs"],
        "summary": "Get all programs",
        "description": "Retrieve a list of all bug bounty programs with optional filtering and pagination",
        "operationId": "getPrograms",
        "security": [{"BearerAuth": []}],
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "description": "Page number for pagination",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "default": 1
            }
          },
          {
            "name": "limit",
            "in": "query",
            "description": "Number of items per page",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 100,
              "default": 20
            }
          },
          {
            "name": "status",
            "in": "query",
            "description": "Filter by program status",
            "required": false,
            "schema": {
              "type": "string",
              "enum": ["draft", "active", "paused", "completed"]
            }
          },
          {
            "name": "search",
            "in": "query",
            "description": "Search term for program name or description",
            "required": false,
            "schema": {
              "type": "string",
              "maxLength": 100
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Programs retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProgramsResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": ["Programs"],
        "summary": "Create a new program",
        "description": "Create a new bug bounty program",
        "operationId": "createProgram",
        "security": [{"BearerAuth": []}],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateProgramRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Program created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Program"
                }
              }
            }
          },
          "400": {
            "description": "Invalid input data",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden - Insufficient permissions",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/programs/{programId}": {
      "get": {
        "tags": ["Programs"],
        "summary": "Get program by ID",
        "description": "Retrieve a specific bug bounty program by ID",
        "operationId": "getProgramById",
        "security": [{"BearerAuth": []}],
        "parameters": [
          {
            "name": "programId",
            "in": "path",
            "description": "Program ID",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Program retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Program"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Program not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": ["Programs"],
        "summary": "Update program",
        "description": "Update an existing bug bounty program",
        "operationId": "updateProgram",
        "security": [{"BearerAuth": []}],
        "parameters": [
          {
            "name": "programId",
            "in": "path",
            "description": "Program ID",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateProgramRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Program updated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Program"
                }
              }
            }
          },
          "400": {
            "description": "Invalid input data",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden - Insufficient permissions",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Program not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": ["Programs"],
        "summary": "Delete program",
        "description": "Delete a bug bounty program",
        "operationId": "deleteProgram",
        "security": [{"BearerAuth": []}],
        "parameters": [
          {
            "name": "programId",
            "in": "path",
            "description": "Program ID",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Program deleted successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden - Insufficient permissions",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Program not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
      "RegisterRequest": {
        "type": "object",
        "required": ["email", "password", "name"],
        "properties": {
          "email": {
            "type": "string",
            "format": "email",
            "description": "User email address"
          },
          "password": {
            "type": "string",
            "minLength": 8,
            "description": "User password (minimum 8 characters)"
          },
          "name": {
            "type": "string",
            "minLength": 2,
            "maxLength": 100,
            "description": "User full name"
          },
          "role": {
            "type": "string",
            "enum": ["user", "researcher"],
            "default": "user",
            "description": "User role"
          }
        }
      },
      "LoginRequest": {
        "type": "object",
        "required": ["email", "password"],
        "properties": {
          "email": {
            "type": "string",
            "format": "email",
            "description": "User email address"
          },
          "password": {
            "type": "string",
            "description": "User password"
          }
        }
      },
      "AuthResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "description": "Operation success status"
          },
          "message": {
            "type": "string",
            "description": "Response message"
          },
          "data": {
            "type": "object",
            "properties": {
              "user": {
                "$ref": "#/components/schemas/UserProfile"
              },
              "token": {
                "type": "string",
                "description": "JWT access token"
              }
            }
          }
        }
      },
      "UserProfile": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "description": "User ID"
          },
          "email": {
            "type": "string",
            "format": "email",
            "description": "User email address"
          },
          "name": {
            "type": "string",
            "description": "User full name"
          },
          "role": {
            "type": "string",
            "enum": ["admin", "researcher", "user"],
            "description": "User role"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time",
            "description": "Account creation timestamp"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time",
            "description": "Last update timestamp"
          }
        }
      },
      "Program": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "description": "Program ID"
          },
          "name": {
            "type": "string",
            "description": "Program name"
          },
          "description": {
            "type": "string",
            "description": "Program description"
          },
          "status": {
            "type": "string",
            "enum": ["draft", "active", "paused", "completed"],
            "description": "Program status"
          },
          "scope": {
            "type": "string",
            "description": "Program scope and targets"
          },
          "rewards": {
            "type": "object",
            "properties": {
              "min": {
                "type": "number",
                "description": "Minimum reward amount"
              },
              "max": {
                "type": "number",
                "description": "Maximum reward amount"
              },
              "currency": {
                "type": "string",
                "description": "Reward currency"
              }
            }
          },
          "rules": {
            "type": "string",
            "description": "Program rules and guidelines"
          },
          "startDate": {
            "type": "string",
            "format": "date-time",
            "description": "Program start date"
          },
          "endDate": {
            "type": "string",
            "format": "date-time",
            "description": "Program end date"
          },
          "createdBy": {
            "type": "string",
            "format": "uuid",
            "description": "User ID of creator"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time",
            "description": "Creation timestamp"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time",
            "description": "Last update timestamp"
          }
        }
      },
      "CreateProgramRequest": {
        "type": "object",
        "required": ["name", "description", "scope"],
        "properties": {
          "name": {
            "type": "string",
            "minLength": 3,
            "maxLength": 100,
            "description": "Program name"
          },
          "description": {
            "type": "string",
            "minLength": 10,
            "maxLength": 1000,
            "description": "Program description"
          },
          "scope": {
            "type": "string",
            "minLength": 10,
            "description": "Program scope and targets"
          },
          "rewards": {
            "type": "object",
            "properties": {
              "min": {
                "type": "number",
                "minimum": 0,
                "description": "Minimum reward amount"
              },
              "max": {
                "type": "number",
                "minimum": 0,
                "description": "Maximum reward amount"
              },
              "currency": {
                "type": "string",
                "default": "USD",
                "description": "Reward currency"
              }
            }
          },
          "rules": {
            "type": "string",
            "description": "Program rules and guidelines"
          },
          "startDate": {
            "type": "string",
            "format": "date-time",
            "description": "Program start date"
          },
          "endDate": {
            "type": "string",
            "format": "date-time",
            "description": "Program end date"
          }
        }
      },
      "UpdateProgramRequest": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "minLength": 3,
            "maxLength": 100,
            "description": "Program name"
          },
          "description": {
            "type": "string",
            "minLength": 10,
            "maxLength": 1000,
            "description": "Program description"
          },
          "status": {
            "type": "string",
            "enum": ["draft", "active", "paused", "completed"],
            "description": "Program status"
          },
          "scope": {
            "type": "string",
            "minLength": 10,
            "description": "Program scope and targets"
          },
          "rewards": {
            "type": "object",
            "properties": {
              "min": {
                "type": "number",
                "minimum": 0,
                "description": "Minimum reward amount"
              },
              "max": {
                "type": "number",
                "minimum": 0,
                "description": "Maximum reward amount"
              },
              "currency": {
                "type": "string",
                "description": "Reward currency"
              }
            }
          },
          "rules": {
            "type": "string",
            "description": "Program rules and guidelines"
          },
          "startDate": {
            "type": "string",
            "format": "date-time",
            "description": "Program start date"
          },
          "endDate": {
            "type": "string",
            "format": "date-time",
            "description": "Program end date"
          }
        }
      },
      "ProgramsResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "description": "Operation success status"
          },
          "data": {
            "type": "object",
            "properties": {
              "programs": {
                "type": "array",
                "items": {
                  "$ref": "#/components/schemas/Program"
                }
              },
              "pagination": {
                "type": "object",
                "properties": {
                  "page": {
                    "type": "integer",
                    "description": "Current page number"
                  },
                  "limit": {
                    "type": "integer",
                    "description": "Items per page"
                  },
                  "total": {
                    "type": "integer",
                    "description": "Total number of items"
                  },
                  "pages": {
                    "type": "integer",
                    "description": "Total number of pages"
                  }
                }
              }
            }
          }
        }
      },
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "example": false
          },
          "message": {
            "type": "string",
            "description": "Error message"
          },
          "errors": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "field": {
                  "type": "string",
                  "description": "Field that caused the error"
                },
                "message": {
                  "type": "string",
                  "description": "Field-specific error message"
                }
              }
            }
          }
        }
      },
      "SuccessResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "example": true
          },
          "message": {
            "type": "string",
            "description": "Success message"
          }
        }
      }
    }
  }
}